+++ core/src/main/java/ai/timefold/solver/core/impl/domain/variable/ExternalizedPreviousElementVariableProcessor.java
@@ -15,7 +15,7 @@ final class ExternalizedPreviousElementVariableProcessor<Solution_>
 
     @Override
     public void setElement(InnerScoreDirector<Solution_, ?> scoreDirector, List<Object> listVariable, Object element,
-            Integer index) {
+            int index) {
         var previous = index == 0 ? null : listVariable.get(index - 1);
         setValue(scoreDirector, element, previous);
     }
